// src/types/car.ts
export type Car = {
  id: string;
  brand: string;
  model: string;
  category: string;
  year: number;
  engine: string;
  power: string;
  transmission: string;
  consumption: string;
  price: number;
  fuel: string;
  colors: string[];
  items: string[];
  description: string;
  image: string;
};